Revert "Refresh open pages after direct MCP mutations (#4081)"
This reverts commit a5686beff6a212c6906efb6c96af0e8d6142a23b.
