--- .changeset/quiet-moles-refresh.md
@@ -1,5 +1,0 @@
----
-"@agent-native/core": patch
----
-
-Refresh open app data after successful mutating actions run through direct MCP tools.
