+++ packages/core/src/action-change-marker.spec.ts
@@ -1,7 +1,6 @@
 import { describe, expect, it } from "vitest";
 
 import {
-  actionChangeDedupeKey,
   actionChangeMarkerValue,
   parseActionChangeMarker,
 } from "./action-change-marker.js";
@@ -20,23 +19,5 @@ describe("action change markers", () => {
       orgId: undefined,
       requestSource: "browser-tab-1",
     });
-  });
-
-  it("uses the marker nonce as a stable fast-path dedupe identity", () => {
-    const target = {
-      actionName: "update-project",
-      owner: "owner@example.com",
-      nonce: "marker-123",
-    };
-
-    expect(actionChangeDedupeKey(target, `action|${target.nonce}`)).toBe(
-      "action|marker-123|update-project|owner@example.com|",
-    );
-    expect(
-      parseActionChangeMarker(
-        "owner@example.com",
-        actionChangeMarkerValue(target),
-      ),
-    ).toEqual(expect.objectContaining({ nonce: "marker-123" }));
   });
 });
