+++ packages/core/src/server/poll.ts
@@ -25,9 +25,7 @@ import { EventEmitter } from "node:events";
 
 import { defineEventHandler, getQuery, setResponseStatus } from "h3";
 
-import { setActionChangeFastPath } from "../action-change-fast-path.js";
 import {
-  actionChangeDedupeKey,
   ACTION_CHANGE_MARKER_KEY,
   parseActionChangeMarker,
   type ActionChangeTarget,
@@ -1359,7 +1357,7 @@ export class AppSyncState {
         },
         dedupeKey !== undefined
           ? {
-              dedupeKey: actionChangeDedupeKey(target, dedupeKey),
+              dedupeKey: `${dedupeKey}|${target.actionName ?? ""}|${target.owner ?? ""}|${target.orgId ?? ""}`,
             }
           : undefined,
       );
@@ -1948,9 +1946,7 @@ export class AppSyncState {
           // marker path.
           this.recordActionChanges(
             [target],
-            target.nonce
-              ? `action|${target.nonce}`
-              : `action|${timestampValue(row.updated_at)}`,
+            `action|${timestampValue(row.updated_at)}`,
           );
         }
         this.lastActionMarkerTs = actionMarkerTs;
@@ -2183,22 +2179,6 @@ export function recordChange(event: {
 }): void {
   getDefaultAppSyncState().recordChange(event);
 }
-
-setActionChangeFastPath((target) => {
-  getDefaultAppSyncState().recordChange(
-    {
-      source: "action",
-      type: "change",
-      key: target.actionName,
-      ...(target.owner ? { owner: target.owner } : {}),
-      ...(target.orgId ? { orgId: target.orgId } : {}),
-      ...(target.requestSource ? { requestSource: target.requestSource } : {}),
-    },
-    target.nonce
-      ? { dedupeKey: actionChangeDedupeKey(target, `action|${target.nonce}`) }
-      : undefined,
-  );
-});
 
 /** Get all changes after a given version. */
 export function getChangesSince(since: number): {
